+++ src/lib/pages/interact-contract/components/query-area/schema-query/SchemaQueryComponent.tsx
